Added more security

+++ api/delete_reservation.php
@@ -4,6 +4,16 @@ require_once "../db.php";
 
 $reservationId = $_POST['deleteId'];
 
+$stmt = $conn->prepare("SELECT id FROM reservations WHERE id = ?");
+$stmt->bind_param('s', $reservationId);
+$stmt->execute();
+$stmt->store_result();
+
+if($stmt->num_rows() == 0){
+    echo json_encode(['status' => 'error', 'message' => 'Невалидно изтриване на несъществуваща резервация!']);
+    exit;
+}
+
 $stmt = $conn->prepare("DELETE FROM reservations WHERE id = ?");
 $stmt->bind_param('s', $reservationId);
